test: 검증 로직 이동에 따라 테스트도 이동
